Move defer to the @kbn/utils package
