+++ src/plugins/bfetch/tsconfig.json
@@ -11,6 +11,7 @@
     "@kbn/config-schema",
     "@kbn/std",
     "@kbn/core-http-server",
+    "@kbn/utils",
   ],
   "exclude": [
     "target/**/*",
