+++ src/plugins/ui_actions/public/service/ui_actions_execution_service.ts
@@ -7,7 +7,7 @@
  */
 
 import { uniqBy } from 'lodash';
-import { defer as createDefer, Defer } from '@kbn/kibana-utils-plugin/public';
+import { defer as createDefer, Defer } from '@kbn/utils';
 import { Action } from '../actions';
 import { buildContextMenuForActions, openContextMenu } from '../context_menu';
 import { Trigger } from '../triggers';
