+++ src/plugins/ui_actions/tsconfig.json
@@ -14,6 +14,7 @@
     "@kbn/i18n",
     "@kbn/es-query",
     "@kbn/ui-actions-browser",
+    "@kbn/utils",
   ],
   "exclude": [
     "target/**/*",
